requirements.txt
accelerate
beartype
einops
environs
gateloop_transformer
jaxtyping
mesh_to_sdf
omegaconf
open3d
opencv-python==4.6.0.66
point_cloud_utils
pytorch_custom_utils
pytorch_lightning
toolz
trimesh
scikit-image
scipy
tyro
x_transformers